Refactored adding/removing parts to grid
Now visiting all part cells and invoke callback to add/remove a part to/from the grid. This enables re-use for a) determine where to place the next part, and b) determine if a part fits there.

Also extended grid cell to maintain the stick orientation. This enables checking that a stick aligns correctly with the hole of the rod in that cell to prevent illegal solutions.

Also extended makeParts to return the total number of cells that the solved puzzle should have. This can be used to quickly check if a puzzle is solved, or if it cannot be solved anymore from the current partial assembly.